--- data/categories.json
@@ -1,16 +1,0 @@
-{
-  "categories": [
-    "증권",
-    "보험",
-    "신탁",
-    "증여 및 상속",
-    "가업승계",
-    "실버케어",
-    "해외이주",
-    "자녀교육",
-    "유언",
-    "세무",
-    "법률",
-    "대출"
-  ]
-}
