--- data/consultingScripts.json
@@ -1,34 +1,0 @@
-{
-  "consultingScripts": [
-    { "sequence": "1", "speaker": "VIP", "content": "안녕하세요." },
-    {
-      "sequence": "2",
-      "speaker": "PB",
-      "content": "안녕하세요. 오랜만이네요! 대출을 원하신다고요~~?"
-    },
-    {
-      "sequence": "3",
-      "speaker": "VIP",
-      "content": "네 용산에 있는 집 있잖아요~ 그거 담보로 대출을 받고 싶어요!"
-    },
-    {
-      "sequence": "4",
-      "speaker": "PB",
-      "content": "네네~ 잠시만요 ~ 아 역 근처에 있는 거 말씀하시는거죠? 그거는 어쩌고~~~~ 그거는 어쩌고~~~~ 그거는 어쩌고~~~~ 그거는 어쩌고~~~~ 그거는 어쩌고~~~~ 그거는 어쩌고~~~~ 그거는 어쩌고~~~~ 그거는 어쩌고~~~~ 그거는 어쩌고~~~~ 그거는 어쩌고~~~~ 그거는 어쩌고~~~~ 그거는 어쩌고~~~~ 그거는 어쩌고~~~~ 그거는 어쩌고~~~~ 그거는 어쩌고~~~~ 그거는 어쩌고~"
-    },
-    {
-      "sequence": "5",
-      "speaker": "VIP",
-      "content": "오늘 저녁에 뭐 드실 예정인가요?"
-    },
-    { "sequence": "6", "speaker": "PB", "content": "아직 계획은 없어요." },
-    { "sequence": "7", "speaker": "VIP", "content": "그럼 같이 식사하실래요?" },
-    { "sequence": "8", "speaker": "PB", "content": "좋아요! 어디로 갈까요?" },
-    {
-      "sequence": "9",
-      "speaker": "VIP",
-      "content": "저녁에 조용한 곳이 좋을 것 같아요."
-    },
-    { "sequence": "10", "speaker": "PB", "content": "그럼 그 근처에서 만나죠!" }
-  ]
-}
