--- data/consultingdata.ts
@@ -1,65 +1,0 @@
-// export type Consulting = {
-//   title: string;
-//   date: string;
-//   status: string;
-// };
-
-// export const consultingData = [
-//   {
-//     title: 'AK프라자 부동산 건 상담요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: 'AK프라자 부동산 건 상담요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-
-//   {
-//     title: 'AK프라자 부동산 건 상담요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: 'AK프라자 부동산 건 상담요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: 'AK프라자 부동산 건 상담요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: 'AK프라자 부동산 건 상담요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: 'AK프라자 부동산 건 상담요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: 'AK프라자 부동산 건 상담요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: 'AK프라자 부동산 건 상담요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: 'AK프라자 부동산 건 상담요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: 'AK프라자 부동산 건 상담요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: 'AK프라자 부동산 건 상담요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: 'AK프라자 부동산 건 상담요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: 'AK프라자 부동산 건 상담요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-// ];
