+++ data/consultingdetaildata.ts
@@ -1,17 +1,17 @@
-export type ConsultingDetail = {
-  category: string;
-  title: string;
-  pb_name: string;
-  customer_name: string;
-  date: string;
-  content: string;
-};
+// export type ConsultingDetail = {
+//   category: string;
+//   title: string;
+//   pb_name: string;
+//   customer_name: string;
+//   date: string;
+//   content: string;
+// };
 
-export const consultingDetailData = {
-  category: '대출',
-  title: 'AK프라자 부동산 건 상담요청합니다.',
-  pb_name: '안유진',
-  customer_name: '조경은',
-  date: '2024.10.22 16:30',
-  content: '내용입니다.',
-};
+// export const consultingDetailData = {
+//   category: '대출',
+//   title: 'AK프라자 부동산 건 상담요청합니다.',
+//   pb_name: '안유진',
+//   customer_name: '조경은',
+//   date: '2024.10.22 16:30',
+//   content: '내용입니다.',
+// };
