--- data/notificationsData.ts
@@ -1,20 +1,0 @@
-// export const notificationsData = [
-//   {
-//     id: 1,
-//     title: '새해 인사 드립니다.',
-//     text: '안녕하세요, 조경은 고객님 안유진 PB입니다. 2025년 새해가 밝았습니다. 새해 복 많이 받으시길 바랍니다~^^',
-//     read: false,
-//   },
-//   {
-//     id: 2,
-//     title: '조경은 고객님을 위한 카드 혜택을 알려드립니다.',
-//     text: '',
-//     read: false,
-//   },
-//   {
-//     id: 3,
-//     title: '2024.04.10일자 상담 일지가 도착했습니다.',
-//     text: '',
-//     read: false,
-//   },
-// ];
