--- data/scheduledata.ts
@@ -1,56 +1,0 @@
-// export type Schedule = {
-//   title: string;
-//   date: string;
-// };
-
-// export const scheduleData = [
-//   {
-//     title:
-//       '건물 매각 건에 대해서 상담 요청합니다.건물 매각 건에 대해서 상담 요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: '건물 매각 건에 대해서 상담 요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: '건물 매각 건에 대해서 상담 요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: '건물 매각 건에 대해서 상담 요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: '건물 매각 건에 대해서 상담 요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: '건물 매각 건에 대해서 상담 요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: '건물 매각 건에 대해서 상담 요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: '건물 매각 건에 대해서 상담 요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: '건물 매각 건에 대해서 상담 요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: '건물 매각 건에 대해서 상담 요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: '건물 매각 건에 대해서 상담 요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-//   {
-//     title: '건물 매각 건에 대해서 상담 요청합니다.',
-//     date: '2024.10.22 16:30',
-//   },
-// ];
